Revert "Feature/authenticator"
