+++ accsys/loginwindow.cpp
@@ -39,11 +39,6 @@ void LoginWindow::on_loginBtn_clicked()
 
 void LoginWindow::on_registerBtn_clicked()
 {
-    if (!registerWindow) {
-        registerWindow = new RegisterWindow(this);
-        connect(registerWindow, &RegisterWindow::destroyed, this, [this]() {
-            registerWindow = nullptr;
-        });
-    }
+    RegisterWindow *registerWindow = new RegisterWindow();
     registerWindow->show();
 }
