+++ accsys/registerwindow.cpp
@@ -1,6 +1,5 @@
 #include "registerwindow.h"
 #include "ui_registerwindow.h"
-#include "HttpClient.h"
 #include <QMessageBox>
 
 RegisterWindow::RegisterWindow(QWidget *parent) :
@@ -14,9 +13,8 @@ RegisterWindow::RegisterWindow(QWidget *parent) :
         QMessageBox::information(this, "Registration", "User registered successfully.");
         this->close();
     });
-    connect(httpClient, &HttpClient::registrationFailure, this, [this](const QString &message) {
-        QMessageBox::information(this, "Registration", "User registered successfully.");
-        this->close();
+    connect(httpClient, &HttpClient::registrationFailure, this, [](const QString &message) {
+        QMessageBox::critical(nullptr, "Registration Error", "Registration failed: " + message);
     });
 }
 
@@ -30,11 +28,5 @@ void RegisterWindow::on_registerButton_clicked()
 {
     QString username = ui->usernameLineEdit->text();
     QString password = ui->passwordLineEdit->text();
-    QString email = ui->emailLineEdit->text();
-    httpClient->registerUser(username, password, email);
-}
-
-void RegisterWindow::on_backButton_clicked()
-{
-    close();
+    httpClient->registerUser(username, password);
 }
