+++ accsys/registerwindow.ui
@@ -7,7 +7,7 @@
     <x>0</x>
     <y>0</y>
     <width>400</width>
-    <height>350</height>
+    <height>300</height>
    </rect>
   </property>
   <property name="sizePolicy">
@@ -25,21 +25,6 @@ font-family: Arial, sans-serif;
 color: white;</string>
   </property>
   <layout class="QVBoxLayout" name="verticalLayout">
-   <item>
-    <widget class="QLabel" name="headerLabel">
-     <property name="text">
-      <string>Register New Account</string>
-     </property>
-     <property name="alignment">
-      <set>Qt::AlignCenter</set>
-     </property>
-     <property name="styleSheet">
-      <string notr="true">font-size: 20px;
-font-weight: bold;
-margin-bottom: 10px;</string>
-     </property>
-    </widget>
-   </item>
    <item>
     <layout class="QFormLayout" name="formLayout">
      <property name="fieldGrowthPolicy">
@@ -62,8 +47,7 @@ margin-bottom: 10px;</string>
 border-radius: 6px;
 background-color: #3E4257;
 font-weight: 600;
-color: white;
-padding: 5px;</string>
+color: white;</string>
        </property>
       </widget>
      </item>
@@ -81,8 +65,7 @@ padding: 5px;</string>
 border-radius: 6px;
 background-color: #3E4257;
 font-weight: 600;
-color: white;
-padding: 5px;</string>
+color: white;</string>
        </property>
        <property name="echoMode">
         <enum>QLineEdit::Password</enum>
@@ -103,8 +86,7 @@ padding: 5px;</string>
 border-radius: 6px;
 background-color: #3E4257;
 font-weight: 600;
-color: white;
-padding: 5px;</string>
+color: white;</string>
        </property>
       </widget>
      </item>
@@ -116,9 +98,7 @@ padding: 5px;</string>
       <string notr="true">border-style: solid;
 border-radius: 6px;
 background-color: #3E4257;
-font-weight: 600;
-padding: 8px;
-margin-top: 10px;</string>
+font-weight: 600;</string>
      </property>
      <property name="text">
       <string>Register</string>
@@ -131,9 +111,7 @@ margin-top: 10px;</string>
       <string notr="true">border-style: solid;
 border-radius: 6px;
 background-color: #3E4257;
-font-weight: 600;
-padding: 8px;
-margin-top: 5px;</string>
+font-weight: 600;</string>
      </property>
      <property name="text">
       <string>Back</string>
