+++ flask/main.py
@@ -4,7 +4,9 @@ import os
 
 app = Flask(__name__)
 
+
 USERS_FILE = 'users.csv'
+
 
 def verify_user(username, password):
     if os.path.exists(USERS_FILE):
@@ -13,31 +15,32 @@ def verify_user(username, password):
         return not user.empty
     return False
 
-def register_user(username, password, email):
+
+def register_user(username, password):
     if os.path.exists(USERS_FILE):
         df = pd.read_csv(USERS_FILE)
         if username in df['username'].values:
             return False
     else:
-        df = pd.DataFrame(columns=['username', 'password', 'email'])
-
-    new_data = pd.DataFrame({'username': [username], 'password': [password], 'email': [email]})
+        df = pd.DataFrame(columns=['username', 'password'])
+    new_data = pd.DataFrame({'username': [username], 'password': [password]})
     df = pd.concat([df, new_data], ignore_index=True)
     df.to_csv(USERS_FILE, index=False)
     return True
+
 
 @app.route('/register', methods=['POST'])
 def register():
     data = request.get_json()
     username = data.get('username')
     password = data.get('password')
-    email = data.get('email')
-    if username and password and email:
-        if register_user(username, password, email):
+    if username and password:
+        if register_user(username, password):
             return jsonify({"status": "success", "message": "User registered successfully."})
         else:
             return jsonify({"status": "error", "message": "User already exists."}), 409
-    return jsonify({"status": "error", "message": "Missing username, password or email."}), 400
+    return jsonify({"status": "error", "message": "Missing username or password."}), 400
+
 
 @app.route('/login', methods=['POST'])
 def login():
@@ -50,6 +53,7 @@ def login():
         else:
             return jsonify({"status": "error", "message": "Invalid username or password."}), 401
     return jsonify({"status": "error", "message": "Missing username or password."}), 400
+
 
 if __name__ == '__main__':
     app.run(debug=True)
