--- registerwindow.ui
@@ -1,125 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>RegisterWindow</class>
- <widget class="QDialog" name="RegisterWindow">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>400</width>
-    <height>300</height>
-   </rect>
-  </property>
-  <property name="sizePolicy">
-   <sizepolicy hsizetype="Preferred" vsizetype="Preferred">
-    <horstretch>0</horstretch>
-    <verstretch>0</verstretch>
-   </sizepolicy>
-  </property>
-  <property name="windowTitle">
-   <string>Register</string>
-  </property>
-  <property name="styleSheet">
-   <string notr="true">background-color: #2a2a3c;
-font-family: Arial, sans-serif;
-color: white;</string>
-  </property>
-  <layout class="QVBoxLayout" name="verticalLayout">
-   <item>
-    <layout class="QFormLayout" name="formLayout">
-     <property name="fieldGrowthPolicy">
-      <enum>QFormLayout::ExpandingFieldsGrow</enum>
-     </property>
-     <property name="rowWrapPolicy">
-      <enum>QFormLayout::WrapAllRows</enum>
-     </property>
-     <item row="0" column="0">
-      <widget class="QLabel" name="usernameLabel">
-       <property name="text">
-        <string>Username:</string>
-       </property>
-      </widget>
-     </item>
-     <item row="0" column="1">
-      <widget class="QLineEdit" name="usernameLineEdit">
-       <property name="styleSheet">
-        <string notr="true">border-style: solid;
-border-radius: 6px;
-background-color: #3E4257;
-font-weight: 600;
-color: white;</string>
-       </property>
-      </widget>
-     </item>
-     <item row="1" column="0">
-      <widget class="QLabel" name="passwordLabel">
-       <property name="text">
-        <string>Password:</string>
-       </property>
-      </widget>
-     </item>
-     <item row="1" column="1">
-      <widget class="QLineEdit" name="passwordLineEdit">
-       <property name="styleSheet">
-        <string notr="true">border-style: solid;
-border-radius: 6px;
-background-color: #3E4257;
-font-weight: 600;
-color: white;</string>
-       </property>
-       <property name="echoMode">
-        <enum>QLineEdit::Password</enum>
-       </property>
-      </widget>
-     </item>
-     <item row="2" column="0">
-      <widget class="QLabel" name="emailLabel">
-       <property name="text">
-        <string>Email:</string>
-       </property>
-      </widget>
-     </item>
-     <item row="2" column="1">
-      <widget class="QLineEdit" name="emailLineEdit">
-       <property name="styleSheet">
-        <string notr="true">border-style: solid;
-border-radius: 6px;
-background-color: #3E4257;
-font-weight: 600;
-color: white;</string>
-       </property>
-      </widget>
-     </item>
-    </layout>
-   </item>
-   <item>
-    <widget class="QPushButton" name="registerButton">
-     <property name="styleSheet">
-      <string notr="true">border-style: solid;
-border-radius: 6px;
-background-color: #3E4257;
-font-weight: 600;</string>
-     </property>
-     <property name="text">
-      <string>Register</string>
-     </property>
-    </widget>
-   </item>
-   <item>
-    <widget class="QPushButton" name="backButton">
-     <property name="styleSheet">
-      <string notr="true">border-style: solid;
-border-radius: 6px;
-background-color: #3E4257;
-font-weight: 600;</string>
-     </property>
-     <property name="text">
-      <string>Back</string>
-     </property>
-    </widget>
-   </item>
-  </layout>
- </widget>
- <resources/>
- <connections/>
-</ui>
